# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dropout Prediction API",
    "description": "Api to predict the dropout",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "Upload Data",
      "description": "For uploading Data"
    },
    {
      "name": "Graph Generation",
      "description": "Generation of different graphs"
    },
    {
      "name": "Train Model",
      "description": "Train the Model using Uploaded Datasets"
    },
    {
      "name": "Prediction",
      "description": "Predict the values using trained model"
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:5000/",
      "description": "Local Development Environment"
    }
  ],
  "paths": {
    "/upload-csv": {
      "post": {
        "tags": [
          "Upload Data"
        ],
        "summary": "Upload a CSV file for training data",
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CSV data uploaded successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "CSV data uploaded successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "No file was uploaded"
                }
              }
            }
          }
        }
      }
    },
    "/distribution-graph": {
      "get": {
        "tags": [
          "Graph Generation"
        ],
        "summary": "Get distribution graph",
        "responses": {
          "200": {
            "description": "Distribution graph generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "png_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "No CSV data available"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Distribution graph not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "An error occurred: error description"
                }
              }
            }
          }
        }
      }
    },
    "/correlation-graph/{selected_column}": {
      "get": {
        "tags": [
          "Graph Generation"
        ],
        "summary": "Get correlation graph",
        "parameters": [
          {
            "name": "selected_column",
            "in": "path",
            "description": "The selected target column for correlation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Correlation graph generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "png_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Correlation graph not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "An error occurred: error description"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "No target column was selected"
                }
              }
            }
          }
        }
      }
    },
    "/training": {
      "get": {
        "tags": [
          "Train Model"
        ],
        "summary": "Get training parameters",
        "responses": {
          "200": {
            "description": "Model training completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accuracy": {
                      "type": "number"
                    },
                    "date_modified": {
                      "type": "string",
                      "format": "date"
                    },
                    "message": {
                      "type": "string"
                    },
                    "precision": {
                      "type": "number"
                    },
                    "training_time": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "File not found:{file_name}"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Failed to save training results to the database: {describe error}"
                }
              }
            }
          }
        }
      }
    },
    "/prediction/multiple-data": {
      "post": {
        "tags": [
          "Prediction"
        ],
        "summary": "Predict on the trained model and update the result in the database for multiple record",
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Model file not found, Please train the model!"
          },
          "NoFileUploadedError": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "No file was uploaded."
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Error parsing CSV data:{describe error}"
                }
              }
            }
          }
        }
      }
    },
    "/prediction/single-data": {
      "post": {
        "tags": [
          "Prediction"
        ],
        "summary": "Predict on the trained model and update the result in the database for single record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "Tuition fees up to date": {
                    "type": "number"
                  },
                  "Age at enrollment": {
                    "type": "number"
                  },
                  "Curricular units 1st sem (approved)": {
                    "type": "number"
                  },
                  "Curricular units 1st sem (grade)": {
                    "type": "number"
                  },
                  "Curricular units 2nd sem (approved)": {
                    "type": "number"
                  },
                  "Curricular units 2nd sem (grade)": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Model file not found, Please train the model!"
          },
          "NoFileUploadedError": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "No file was uploaded."
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Error parsing CSV data:{describe error}"
                }
              }
            }
          }
        }
      }
    }
  }
}